Various Bug Fixes for the Autoscheduler - LOOPer branch

+++ benchmarks/autoscheduler_benchmarks/polybench/LARGE/function_deriche_LARGE/function_deriche_LARGE_wrapper.cpp
@@ -12,13 +12,13 @@ using namespace std;
 int main(int, char **argv)
 {
 
-	double *b_A = (double*)malloc(4096*2160*sizeof(double));
-	parallel_init_buffer(b_A, 4096*2160, (double)19);
-	Halide::Buffer<double> buf01(b_A, 4096, 2160);
+	float *b_A = (float*)malloc(4096*2160*sizeof(float));
+	parallel_init_buffer(b_A, 4096*2160, (float)19);
+	Halide::Buffer<float> buf01(b_A, 4096, 2160);
 
-	double *b_x = (double*)malloc(4096*2160* sizeof(double));
-	parallel_init_buffer(b_x, 4096*2160, (double)36);
-	Halide::Buffer<double> buf02(b_x, 4096, 2160);
+	float *b_x = (float*)malloc(4096*2160* sizeof(float));
+	parallel_init_buffer(b_x, 4096*2160, (float)36);
+	Halide::Buffer<float> buf02(b_x, 4096, 2160);
