+++ benchmarks/autoscheduler_benchmarks/polybench/MEDIUM/function_deriche_MEDIUM/function_deriche_MEDIUM_wrapper.cpp
@@ -12,13 +12,13 @@ using namespace std;
 int main(int, char **argv)
 {
 
-	double *b_A = (double*)malloc(720*480*sizeof(double));
-	parallel_init_buffer(b_A, 720*480, (double)19);
-	Halide::Buffer<double> buf01(b_A, 720, 480);
+	float *b_A = (float*)malloc(720*480*sizeof(float));
+	parallel_init_buffer(b_A, 720*480, (float)19);
+	Halide::Buffer<float> buf01(b_A, 720, 480);
 
-	double *b_x = (double*)malloc(720*480* sizeof(double));
-	parallel_init_buffer(b_x, 720*480, (double)36);
-	Halide::Buffer<double> buf02(b_x, 720, 480);
+	float *b_x = (float*)malloc(720*480* sizeof(float));
+	parallel_init_buffer(b_x, 720*480, (float)36);
+	Halide::Buffer<float> buf02(b_x, 720, 480);
