+++ benchmarks/autoscheduler_benchmarks/polybench/SMALL/function_floyd_warshall_SMALL/function_floyd_warshall_SMALL_wrapper.cpp
@@ -12,9 +12,9 @@ using namespace std;
 int main(int, char **argv)
 {
 
-	double *b_D = (double*)malloc(180*180*sizeof(double));
-	parallel_init_buffer(b_D, 180*180, (double)19);
-	Halide::Buffer<double> buf00(b_D, 180, 180);
+	int *b_D = (int*)malloc(180*180*sizeof(int));
+	parallel_init_buffer(b_D, 180*180, (int)19);
+	Halide::Buffer<int> buf00(b_D, 180, 180);
